requirements.txt
git+https://github.com/revoltchat/revolt.py
discord.py
python-dotenv
aenum
ulid
typing-extensions
nest-asyncio
